requirements.txt
flask==3.0.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
qrcode[pil]==7.4.2
pillow==10.3.0
openpyxl==3.1.2
werkzeug==3.0.3
python-dotenv==1.0.1
sqlalchemy==2.0.30